+++ src/functionsfactory.ts
@@ -246,3 +246,12 @@ function currentDate() {
   return new Date();
 }
 FunctionFactory.Instance.register("currentDate", currentDate);
+
+function today(params: any[]) {
+  var res = new Date();
+  if (Array.isArray(params) && params.length == 1) {
+    res.setDate(res.getDate() + params[0]);
+  }
+  return res;
+}
+FunctionFactory.Instance.register("today", today);