+++ src/questionnonvalue.ts
@@ -24,6 +24,9 @@ export class QuestionNonValue extends Question {
   public get hasComment(): boolean {
     return false;
   }
+  public hasErrors(fireCallback: boolean = true, rec: any = null): boolean {
+    return false;
+  }
   public getAllErrors(): Array<SurveyError> {
     return [];
   }